Completeness of vaccination dates

Of the doses documented on a card, how many carry a date, and how many carry a date complete enough to compute age at vaccination.

date_completeness(x, by = NULL)

Arguments

x

A vcs_data object.

by

Grouping variables, drawn from the child-level table.

Value

A tibble with n_documented, n_dated, prop_dated, n_partial and prop_partial per group.

Examples

date_completeness(vcs_example)
#> # A tibble: 1 × 5
#>   n_documented n_dated n_partial prop_dated prop_partial
#>          <int>   <int>     <int>      <dbl>        <dbl>
#> 1         1346    1161       104      0.863       0.0773
date_completeness(vcs_example, by = ~vaccine)
#> # A tibble: 14 × 6
#>    n_documented n_dated n_partial prop_dated prop_partial vaccine
#>           <int>   <int>     <int>      <dbl>        <dbl> <chr>  
#>  1          137     116        13      0.847      0.0949  BCG    
#>  2           94      80         7      0.851      0.0745  IPV1   
#>  3           92      75         8      0.815      0.0870  MCV1   
#>  4          102      88         7      0.863      0.0686  OPV0   
#>  5           81      68         8      0.840      0.0988  OPV1   
#>  6           69      61         5      0.884      0.0725  OPV2   
#>  7           49      47         1      0.959      0.0204  OPV3   
#>  8          131     114        11      0.870      0.0840  PCV1   
#>  9          104      99         1      0.952      0.00962 PCV2   
#> 10           74      62         4      0.838      0.0541  PCV3   
#> 11          136     111        16      0.816      0.118   PENTA1 
#> 12          110      96         8      0.873      0.0727  PENTA2 
#> 13           81      66        10      0.815      0.123   PENTA3 
#> 14           86      78         5      0.907      0.0581  YF