Intervals between consecutive dated doses

For every dose with a previous_dose in the schedule, the number of days between the two doses when both carry usable card dates. Nothing is computed for a pair with a missing or partial date.

derive_dose_intervals(x, schedule = NULL)

Arguments

x

A vcs_data object.

schedule

A vcs_schedule(); defaults to the one stored in x.

Value

A tibble with child_id, antigen, from, to, interval_days, scheduled_interval_days and the child-level columns of x.

See also

Examples

head(derive_dose_intervals(vcs_example))
#> # A tibble: 6 × 37
#>   child_id antigen from  to    interval_days scheduled_interval_d…¹ household_id
#>   <chr>    <chr>   <chr> <chr>         <dbl>                  <dbl> <chr>       
#> 1 CH00051  OPV     OPV0  OPV1             33                     28 uuid:HH0005 
#> 2 CH00121  OPV     OPV0  OPV1             56                     28 uuid:HH0012 
#> 3 CH00221  OPV     OPV0  OPV1             50                     28 uuid:HH0022 
#> 4 CH00551  OPV     OPV0  OPV1             69                     28 uuid:HH0055 
#> 5 CH00981  OPV     OPV0  OPV1             29                     28 uuid:HH0098 
#> 6 CH01071  OPV     OPV0  OPV1             54                     28 uuid:HH0107 
#> # ℹ abbreviated name: ¹​scheduled_interval_days
#> # ℹ 30 more variables: interview_id <chr>, caregiver_index <int>,
#> #   child_index <int>, weight <dbl>, child_dob <date>, sex <chr>,
#> #   card_seen <lgl>, child_name <chr>, caregiver_name <chr>, psu <chr>,
#> #   segment <chr>, stratum <chr>, province <chr>, district <chr>,
#> #   health_zone <chr>, health_area <chr>, residence <chr>,
#> #   interview_date <date>, duration_min <dbl>, enumerator <chr>, team <chr>, …