Dictionary for a KC v9 SurveyCTO wide export

Dictionary for a KC v9 SurveyCTO wide export

kc9_dictionary(
  n_caregivers = 3L,
  n_children = 3L,
  psu = "PVT06",
  health_zone = "PVT05",
  weight = NULL,
  stratum = NULL,
  gps_suffix = c("-Latitude", "-Longitude", "-Accuracy"),
  date_formats = c("%m/%d/%Y", "%Y-%m-%d", "%d/%m/%Y"),
  roster = FALSE
)

Arguments

n_caregivers, n_children

Repeat extents present in the export.

psu

Source column holding the PSU; "PVT06" is populated on every row while submission_psu is not.

health_zone

Source column holding the health zone.

weight

Optional source column holding the child-level weight (a column you have joined onto the export yourself).

stratum

Optional source column holding the sampling stratum.

gps_suffix

How the export splits the geopoint: SurveyCTO desktop exports write PVT_GPS-Latitude; some server exports write PVT_GPS_latitude.

date_formats

Date formats to try, most common first. Desktop exports write m/d/Y.

roster

Map the child level onto the rost_* columns built by kc9_prepare() with unit = "roster", so that every child listed in the household listing becomes a row. n_children is then the number of roster slots and n_caregivers is 1.

Value

A vcs_dictionary().

See also

Examples

kc9_dictionary()
#> <vcs_dictionary>
#>   repeats: 3 caregiver(s) x 3 child(ren)
#>   household level:
#>     interview_id   -> KEY
#>     household_id   -> KEY
#>     psu            -> PVT06
#>     segment        -> PVT07
#>     health_zone    -> PVT05
#>     residence      -> HHI07
#>     interview_date -> today
#>     duration_min   -> duration
#>     enumerator     -> PVT02
#>     team           -> PVT03
#>     supervisor     -> PVT04
#>     gps_lat        -> PVT_GPS-Latitude
#>     gps_lon        -> PVT_GPS-Longitude
#>     gps_accuracy   -> PVT_GPS-Accuracy
#>     eligible       -> ELHH
#>     consent        -> HHI09
#>     n_eligible     -> RHH04
#>     telephone      -> RHH02_phone
#>     address        -> HHI06
#>   child level:
#>     caregiver_id   -> EC04_{c}_{k}
#>     child_dob      -> child_dob_{c}_{k}
#>     age_months     -> child_months_{c}_{k}
#>     sex            -> child_sex_{c}_{k}
#>     card_seen      -> CVH_card_shown_{c}_{k}
#>     child_name     -> EC03_{c}_{k}
#>     caregiver_name -> CB02_{c}
#>   vaccine level:
#>     card_status   -> CVH{vv}_{c}_{k}
#>     card_date     -> CVH{vv}_date_{c}_{k}
#>     recall_status -> recall_status_{vaccine}_{c}_{k}
#>     recall_ever   -> recall_ever_{antigen}_{c}_{k}
#>     recall_count  -> recall_count_{antigen}_{c}_{k}